services/agent-swarm-runtime/pyproject.toml
[project]
name = "daes-agent-swarm-runtime"
version = "1.0.0"
description = "Deterministic 1000-agent swarm runtime for DAES."
requires-python = ">=3.12"
dependencies = [
  "fastapi==0.115.5",
  "uvicorn[standard]==0.32.1",
  "numpy==2.1.3",
  "pyyaml==6.0.2",
  "weaviate-client==4.9.4",
  "pydantic==2.9.2",
  "httpx==0.27.2",
  "prometheus-client==0.21.0",
  "eth-hash[pycryptodome]==0.7.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
